extract listenResize to a common util function
